Move struct integration tests into structs module
